big system updates 5
working certificate category and css updates

+++ database/migrations/2024_11_10_220731_create_certificates_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('title');
             $table->string('organization')->nullable();
             $table->string('designation')->nullable();
-            $table->integer('days')->nullable();
+            $table->float('days')->nullable();
             $table->string('date');
             $table->text('raw_text');
             $table->float('points')->nullable();
